+++ src/commands/features.ts
@@ -3,78 +3,64 @@ import { workspace } from 'vscode';
 import * as gitflowUtils from '../helpers/gitflowUtils';
 import * as gitUtils from '../helpers/gitUtils';
 import * as path from 'path';
-import { ConfigSettings } from '../settings/configSettings';
+import { InitConfigSettings } from '../settings/configSettings';
 import { BranchSetting } from '../settings/branchSettings';
 
 const config = workspace.getConfiguration(); 
-const configValues = config.get('gitflow4code.init') as ConfigSettings;
+const initValues = config.get('gitflow4code.init') as InitConfigSettings;
+const askForDeletion = config.get('gitflow4code.askBeforeDeletion') as Boolean;
+const deleteByDefault = config.get('gitflow4code.deleteBranchByDefault') as Boolean;
 
-export function run(outChannel) {
-    var itemPickList = [
+export function run(outChannel, action) {
+    if(action === 'start') {
+        var itemPickList = [
             { 
-                label: 'Start Feature from ' + configValues.develop,
+                label: 'Start Feature from ' + initValues.develop,
                 description: ''
             },
             { 
                 label: 'Start Feature from another base branch',
                 description: ''
-            },
-            {
-                label: 'Finish Feature',
-                description: ''
             }
         ];
-    vscode.window.showQuickPick(itemPickList).then(function(item) {
-        if(!item) return;
         
-        outChannel.clear();
-        if(item.label === itemPickList[0].label) 
-            vscode.window.showInputBox({ prompt: 'Name of Feature: ' }).then(val => startFeature(outChannel, val, configValues.develop));
-        else if(item.label === itemPickList[1].label)
-            getBranchNames(outChannel, item.label);
-        else
-            finishFeature(outChannel);
-        
-    });
-}
-
-function getBranchNames(outChannel, branchName) {
-    var branchList = gitUtils.getGitRepositoryPath(vscode.workspace.rootPath).then(function (gitRepositoryPath) {
-        gitUtils.getBranchList(gitRepositoryPath).then((branches) => {
-
-            var branchList = branches as string[];
+        vscode.window.showQuickPick(itemPickList).then(function(item) {
+            if(!item) return;
             
-            var filteredBranchList = branchList.map((value) => {
-                // if(value.replace('*', '').trim() === configValues.develop || value.replace('*', '').trim().startsWith(configValues.features))
-                return value.replace('*', '').trim();
-            }).filter(x => !!x);
-        
-            var branchPickList = [];
-            filteredBranchList.forEach(branchName => {
-                if(branchName === configValues.develop)
-                    branchPickList.push( { label: configValues.develop, description: 'create feature branch using ' + configValues.develop + ' as your base'});
-                else
-                branchPickList.push( { label: branchName, description: 'create feature branch using ' + branchName + ' as your base'});
+            outChannel.clear();
+            if(item.label === itemPickList[0].label) 
+                vscode.window.showInputBox({ prompt: 'Name of Feature: ' }).then(val => startFeature(outChannel, val, initValues.develop));
+            else {
+                gitUtils.getBranchList(workspace.rootPath).then((features) => {
+                    var branchPickList = [];
+                    features.forEach(branchName => {
+                        if(branchName === initValues.develop)
+                            branchPickList.push( { label: initValues.develop, description: 'create feature branch using ' + initValues.develop + ' as your base'});
+                        else
+                            branchPickList.push( { label: branchName, description: 'create feature branch using ' + branchName + ' as your base'});
+                    });
+                
+                    vscode.window.showQuickPick(branchPickList).then(function(item) {
+                        if(!item) return;
+                
+                        outChannel.clear();
+                        vscode.window.showInputBox({ prompt: 'Name of Feature: ' }).then(val => startFeature(outChannel, val, item.label));
+                    });
+                });
+            }
+        });
+    }
+    else if (action === 'finish') {
+        if(askForDeletion)
+            vscode.window.showInputBox({ prompt: 'Would you like this feature branch deleted after finishing? (y/n)' }).then(function(val) {
+                if(val !== undefined && (val.toLowerCase() === 'y' ||  val.toLowerCase() === 'n')) { 
+                    var deleteBranch = val.toLowerCase() === 'y';
+                    finishFeature(outChannel, deleteBranch); 
+                }
             });
-        
-            vscode.window.showQuickPick(branchPickList).then(function(item) {
-                if(!item) return;
-        
-                outChannel.clear();
-                vscode.window.showInputBox({ prompt: 'Name of Feature: ' }).then(val => startFeature(outChannel, val, item.label));
-            });
-        }, genericErrorHandler);
-    });
-
-    function genericErrorHandler(error) {
-        if(error.code && error.syscall && error.code === 'ENOENT' && error.syscall === 'spawn git')
-            vscode.window.showErrorMessage('Cannot find git installation');
-        else {
-            outChannel.appendLine(error);
-            outChannel.show();
-            vscode.window.showErrorMessage('There was an error, please view details in output log');
-        }
-    } 
+        else 
+            finishFeature(outChannel, deleteByDefault);
+    }
 }
 
 function startFeature(outChannel, featureName, baseBranch) {
@@ -95,7 +81,7 @@ function startFeature(outChannel, featureName, baseBranch) {
         }
 
         let featuresConfig = config.get('gitflow4code.features') as BranchSetting[];
-        featuresConfig.push(new BranchSetting(configValues.features + featureName, baseBranch));
+        featuresConfig.push(new BranchSetting(initValues.features + featureName, baseBranch));
         config.update('gitflow4code.features', featuresConfig);
         
         outChannel.append(log);
@@ -112,25 +98,27 @@ function startFeature(outChannel, featureName, baseBranch) {
     } 
 }
 
-function finishFeature(outChannel) {
+function finishFeature(outChannel, deleteBranch) {
     gitUtils.getGitRepositoryPath(vscode.workspace.rootPath).then(function (gitRepositoryPath) {
         gitUtils.getCurrentBranchName(vscode.workspace.rootPath).then((branchName) => {
-            if(branchName.toString().startsWith(configValues.features)) {
+            if(branchName.toString().startsWith(initValues.features)) {
                 let featuresConfig = config.get('gitflow4code.features') as BranchSetting[];
                 let featureSetting = featuresConfig.find((feature) => feature.name === branchName.toString());
                 if(!featureSetting) 
-                    featureSetting = new BranchSetting(branchName.toString(), configValues.develop);
+                    featureSetting = new BranchSetting(branchName.toString(), initValues.develop);
 
-                gitflowUtils.finishFeature(gitRepositoryPath, branchName.toString(), featureSetting.base).then(finishFeature, genericErrorHandler);
+                gitflowUtils.finishFeature(gitRepositoryPath, branchName.toString(), featureSetting.base, deleteBranch).then(finishFeature, genericErrorHandler);
                 function finishFeature(log) {
                     if(log.length === 0) {
                         vscode.window.showInformationMessage('Nothing to show');
                         return;
                     }
 
-                    let featureIndex = featuresConfig.indexOf(featureSetting);
-                    featuresConfig.splice(featureIndex, 1);
-                    config.update('gitflow4code.features', featuresConfig);
+                    if(deleteBranch) {
+                        let featureIndex = featuresConfig.indexOf(featureSetting);
+                        featuresConfig.splice(featureIndex, 1);
+                        config.update('gitflow4code.features', featuresConfig);
+                    }
                     
                     outChannel.append(log);
                     outChannel.show();