+++ test/features.test.ts
@@ -1,0 +1,17 @@
+//
+// Note: This example test is leveraging the Mocha test framework.
+// Please refer to their documentation on https://mochajs.org/ for help.
+//
+
+// The module 'assert' provides assertion methods from node
+import * as assert from 'assert';
+import * as features from '../src/commands/features';
+
+// Defines a Mocha test suite to group tests of similar kind together
+suite("Features Tests", function () {
+
+    // Defines a Mocha unit test
+    test("Show picklist for features", function() {
+        // features.run(null);
+    });
+});