+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - 2018-4-25
+### Added
+- New statusbar button to quickly finish features, releases, or hotfixes
+- New keyboard shortcuts for quickly running commands
+- New configuration value, askBeforeDeletion. When finishing a feature, release, and/or hotfix, you will now be asked if you would like to delete the branch (only local not remote if you have pushed it). Setting is true by default. If you set to false, it will follow the behavior of the deleteBranch setting.
+- New configuration value, deleteBranchByDefault. When finishing a feature, release, and/or hotfix, this setting will determine if the branch is deleted by default or not. Setting is false by default. If you set to true, and askBeforeDeletion is set to false, then branches will be deleted without question.
+### Changed
+- In support of the new statusbar button, the command menu structure had to be changed somewhat. Starting & Finishing features, releases, and/or hotfixes commands are now listed on the command menu (Cmd+P/Ctrl+P), as are Initialize and Git Status.
+
 ## [1.1.0] - 2018-4-16
 ### Added
 - Ability to create release and/or hotfix branches based on branches other than default branch ("development" and "production" respectively)